requirements.txt
django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
pydantic==2.5.0
openai==1.3.0
gunicorn==20.1.0
python-dotenv==1.0.0
pydantic-ai==0.0.14
